requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1625643473416/work
certifi==2021.10.8
dj-database-url==0.5.0
Django @ file:///tmp/build/80754af9/django_1625585912945/work
django-heroku==0.3.1
django-rest-auth==0.9.5
djangorestframework==3.13.1
gunicorn==20.1.0
Pillow==9.0.1
psycopg2 @ file:///C:/ci/psycopg2_1612298803839/work
pytz==2021.3
six==1.16.0
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
whitenoise==6.0.0
wincertstore==0.2
